+++ NET1814_MilkShop.Repositories/Models/CheckoutModels/CheckoutResponseModel.cs
@@ -5,8 +5,8 @@ public class CheckoutResponseModel
     public Guid? OrderId { get; set; }
     public Guid? CustomerId { get; set; }
     public string? FullName { get; set; }
-    public decimal TotalAmount { get; set; }
-    public decimal ShippingFee { get; set; }
+    public int TotalAmount { get; set; }
+    public int ShippingFee { get; set; }
     public string? Address { get; set; }
     public string? PhoneNumber { get; set; }
     public string? Note { get; set; }
