+++ NET1814_MilkShop.Repositories/Models/ProductModels/UpdateProductModel.cs
@@ -9,9 +9,9 @@ namespace NET1814_MilkShop.Repositories.Models.ProductModels
         [Range(0, int.MaxValue, ErrorMessage = "Quantity must be greater than 0")]
         public int? Quantity { get; set; }
         [Range(0, double.MaxValue, ErrorMessage = "Original Price must be greater than 0")]
-        public decimal? OriginalPrice { get; set; }
+        public int? OriginalPrice { get; set; }
         [Range(0, double.MaxValue, ErrorMessage = "Sale Price must be greater than 0")]
-        public decimal? SalePrice { get; set; }
+        public int? SalePrice { get; set; }
         public string? Thumbnail { get; set; }
         [Range(1, int.MaxValue, ErrorMessage = "Category must be greater than 1")]
         public int? CategoryId { get; set; }
